Extract `LazyAdapter`
